Fix PTSampler.run_mcmc()
The generic Sampler.run_mcmc() function wouldn't work right for PTSampler
because it passed the RNG state, `rstate0`, as the second positional argument
to sample(), while for PTSampler the second positional argument was the extra
`lnlike0` parameter. If you tried to pass `lnlike0` yourself, then, you'd
get an exception for a doubly-specified keyword.

Anyway, the best solution is probably to pass rstate0 as a keyword argument so
that subclasses can have differently-ordered arguments if they want. Do that.
This also requires fixing a discrepancy between the documentation and
implemention of the M-H sampler, which claimed the RNG state keyword was
`rstate0` when it was really being called `randomstate`.